requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.49
pydantic[email]==2.5.2
pydantic-settings==2.7.0
bcrypt==5.0.0
PyJWT==2.12.1
python-multipart==0.0.6
alembic==1.12.1